requirements.txt
python==3.8.13
pandas
ipykernel
pystac
pystac-client
odc-stac
planetary-computer
netCDF4
matplotlib
rioxarray
xarray-spatial
Jinja2
stackstac
pandarallel
ipywidgets
load-dotenv